Rename chemistry FTQC helpers
